[FIX] project_enterprise: remove spreadsheet serivces in project sharing
Before this commit, when the collaborator uses the search view to add
group by or use favorite items, he has a traceback because of the
Service spreadsheet is not available.

This commit removes this service to avoid to use the spreadsheet in
project sharing feature.

task-2633229

Community PR: odoo/odoo#/77156
closes #21155

X-original-commit: f66c1723d85c93022d7a0f9ba2a8a4634e343ae1