[IMP] timesheet_grid, web_grid: default date when adding a line
Currently, the date set by default is the current day of the week
but on the period displayed.

step to reproduce:-
go to timesheets > my timesheets >
go to another week than the current one > add a line

After this commit, except for the current week, the date set by default
should be the first working day of the period displayed.

Task-Id:2585241

closes odoo/enterprise#21538

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>