[IMP] timesheet_grid, web_grid: add the 'day' mode in the my timesheet grid
Currently, week and month mode is available in my timesheet grid, now day
has been added, so that the user can easily view the timesheet of a
particular day.

Moreover, when the user is in mobile view the `day` range is duplicated
since we add this range in the grid view. Indeed, by default when the
user is in mobile view of grid view, we add by default the `day` range
without checking if the xml view contains this range.

This commit first extracts the ranges contained in the xml grid view and
then add the `day` range as first element of the ranges list if the user
is in mobile view and this range is not in the ranges list yet.

Closes odoo/enterprise#21692
task-2652210