[FIX] documents_spreadsheet: translate terms in o-spreadshete templates
Terms in o-spreadsheet templates were not translated because the
translation function from odoo was not correctly set in the o-spreadsheet
env. Hence the default translation function of o-spreadsheet (which does
nothing) was used instead.

Note: terms were correctly extracted, so there are available translations.

Also fix the template dialog title which was not translatable and update
the .pot file. It seems that previous changes forgot to do it.

closes odoo/enterprise#21592

X-original-commit: 03180daa8dbaf1a123485476a45bf67a0a596390
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>