[IMP] sign: allows linking sign items with x2many fields
In sign it is possible to link sign items to partner fields by setting the
auto_field field in sign.item.type. However when linking x2many fields, a
traceback occurs since it tries to access one field on a recordset with many
records. This PR prevents the traceback by filtering the highest priority
record from the recordset (defined by active and id values as defined in the
spec) and using it to return the property.

task-2658338

Part-of: odoo/enterprise#21541