[FIX] l10n_cl_edi_stock: Handle unexpected responses and invalid token codes
This fix is exactly the same as the one that was done for l10n_cl_edi in
https://github.com/odoo/enterprise/pull/21503 and https://github.com/odoo/enterprise/pull/20868 .
The logic is duplicated between modules and affected by the same issues.

The UnexpectedXMLResponse exception which was added in l10n_cl_edi was not
yet caught in l10n_cl_edi_stock, so if the condition was raised it would
result in a traceback in the UI. This exception is caught the same way as
in l10n_cl_edi now.

Because the token being invalid was only checked after the response had been
analyzed, and unexpected responses cause the function to return and keep the
SII status the same, the invalid token would not get cleared. Like for
l10n_cl_edi, checking for an invalid token response before analyzing the response
and clearing the token if it was invalid fixes this. No retry was implemented
like in l10n_cl_edi, so if this happens the user (or cron job) needs to verify twice
and after that the SII status will update.

X-original-commit: ca3e3ac54b8c3e06d580ecb85792debb5b882abd
Part-of: odoo/enterprise#22391