[FIX] industry_fsm: make the done as closed stage
Currently, when mark the task as done it moves the task to the canceled
stage instead of done because the done stage is not set as close stage
in data.

So in this commit, make the done stage as is_close true so when marking
the task as done is should move to the done stage instead of cancel.

Closes: https://github.com/odoo/enterprise/pull/21602
Task-2665826

X-original-commit: cc2726ebba04cdbda31f2b0a627c6fc3ee917fff
Part-of: odoo/enterprise#22011