[FIX] industry_fsm: add attr copy false on fsm_done
Currently, when duplicate the done task it will not show the
all the button on task because it still copy the fsm_done and already
mark the new task as done.

So in this commit, make the fsm_done as copy False so task should not be
marked as done on duplication of record.

Closes: https://github.com/odoo/enterprise/pull/21602
Task-2665826

closes odoo/enterprise#22011

X-original-commit: 56dfa09832aea015dca64457857e485ff00f28db
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>