[FIX] {sale_}project_forecast: update project if the one in task changed
Before this commit, when a task is linked to a shift and this shift is
linked to a sales order line. The project of this shift is not changed
if a project is already set, because its compute is override in
`sale_project_forecast` to give the one in sales order line as default
value. However, we need to recompute if the project in the shift is
already set to ensure the project is the one of the task linked.
Therefore, when the user changes the project of
the task, the shift has not the new project of this task but the old one
of this task.

This commit changes the `_compute_project_id` method in the `planning.slot`
model to recompute the project of the shifts if the project is already
set in those shifts or if we don't any project based on the sales order
line linked.

task-2665899

Community PR: #79034

closes #21907

X-original-commit: 3aa6f93
Part-of: odoo/enterprise#22056