[IMP] social: raise an error when account not found while adding stream
Before this commit,
While adding the streams, if the account could not be found with the
provided credential, it was not properly handled (user was not able
to understand what happend as sometimes user redirect back to empty
dashboard or faced internal server error).

After this commit,
It will show proper error with more information if no account could
be added so that user knows what happened.

task-2622803

closes odoo/enterprise#22019

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>