[FIX] account_{asset,reports}: sorting in report with hierarchy
Since f3e163c9653e34769cc0860fa45e62756c2820f3, the lines are sorted in
the hierarchy.
The issue is that for the sorting to work, the account's code need to
part of the name string.
We can now specify the code in models where it isn't part of the
displayed string so that the sorting can be made.

closes odoo/enterprise#21601

X-original-commit: dd0b9fcc0f20ca42bbb83eff7ccd6cb44392a3df
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>