[IMP] hr_payroll: improve scrollbar in edit payslip lines
The scrollbar in the edit payslip lines wizard could be quite unusable
in some cases.
On large displays, if the payslip lines takes too much space both
vertically and horizontally, both scrollbars would be separated from
eachother.
This commit makes it so that both scrollbars are on the same element in
the view.

TaskId-2659268

closes odoo/enterprise#21593

Signed-off-by: Kevin Baptiste <kba@odoo.com>