[FIX] account_invoice_extract: fix required fields check
Because we are using the Form simulation tool to fill the invoice,
it will check by default that the required fields of the form view
are all filled.
This is annoying as the OCR will leave these blank and the user will
get a traceback.
The issue is solved by removing the check of required fields before
saving the form.

closes odoo/enterprise#21821

X-original-commit: 196fd82f593066b8ce32ddf6c2e4d6d57b90003f
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux <lobaudoux@users.noreply.github.com>