[FIX] appointment: hide default bootstrap caret on drop-down toggle
Currently when appointment is installed, from the back-end of calendar
app we can share the availabilities(see sidebar at the right). This
sharing drop-down toggle uses a custom icon, and so we need to hide
the default one that is provided by bootstrap, otherwise it displays
custom icon and default caret both, and ruins the UI.

To hide the default caret, we can either remove the `dropdown-toggle`
class from button, or we can simply hide the default caret. The first
solution might break x-path, and it can also break the feature(if there
are multi level drop-down menus).

This commit simply hides the default caret and thus only shows the
custom icon for sharing drop-down toggle.

TaskID-2666793

closes odoo/enterprise#21777

X-original-commit: e661092572870ff25f9ac96d7eff4120931202a3
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>