[IMP] website_helpdesk, *: add helpdesk to the website configurator
*: website_enterprise

Currently, when creating and configuring a website, there is no option
to select the Helpdesk feature.

After this commit, when creating a website, allow the creator to select
the "Helpdesk" option from the offered features, which would install the
Helpdesk application and create a functional "Help" page where tickets
can be submitted.

task-2666493

closes odoo/enterprise#21632

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>