[FIX] documents_spreadsheet: don't mutate pivot formula on the network
On runbot
1. Insert the demo CRM lead pivot in a spreadsheet
=> F1 is a Total formula
2. Reload the spreadsheet
=> F2 has magically become a measure formula

the command is mutated in place :(
It's the mutated commands which is send over the network and becomes the
accepted revision

I would like to make the dispatched commands immutable (in o-spreadsheet),
but I didn't find a proper way (yet). A simple deep copy with JSON stringify
and parse would probably do it, but there's currently one core command
which is not serializable (the command is only internal, never sent on the
network). In the mean time, this commit is a quick fix. See task 2667116

Task 2662680

closes odoo/enterprise#21830

X-original-commit: 4a25fde9d03f3923d89096ea862539719caa2734
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>