[FIX] helpdesk: fix format of numbers in dashboard
Before this commit, the number displayed could have more than 2 digits
and the float numbers represent the number of hours are not formatted as
time.

This commit formats the numbers displayed in the dashboard to show at
most 2 digits and format in time for the ones representing the number of
hours.

task-2665899

Community PR: #79034

closes #21907

X-original-commit: 980bd29102e02e4c37d774e5b793ff5c8b85f78a
Part-of: odoo/enterprise#22056