[FIX] web_mobile: mobile routes' tests should be closer to app's usage
This commit fixes the matching for the wrong database name during
authentification from the mobile app to use the same pattern as the
app's codebase in the error message, instead of verifying the equality
of the whole message.

Note: this was spotted on the runbot after an upgrade of the `pq`
library which prefixed this kind of error messages.

Before:
> FATAL:  database "mydatabase" does not exist

After:
> connection to server on socket "/var/run/postgresql/.s.PGSQL.5432" failed: FATAL:  database "mydatabase" does not exist

Thanks MOC for the report.

closes odoo/enterprise#21682

X-original-commit: 70b21a024eecfbff6f5ef2b51895826a1769633a
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>