[FIX] account_inter_company_rules: fix intercompany billing flow
Currently, If we create a SO and create a draft invoice with the customer as another
company whose rule_type for the inter-company transaction is selected
to 'invoice_and_refund'. An error message is thrown as Incompatible companies on
records on confirming the invoice.
NB: It can also more likely be a configuration issue if journals are not
correctly created.

so in this commit as a simple fix, we intend to set the company_id of the analytic
account as 'false' in the respective SO (as well as linked project ) to continue
the billing  flow.

task-2587103

closes odoo/enterprise#22446

X-original-commit: e0232fb7845f36ae0095593bb5f66f9b1211ce34
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>