[FIX]account_reports:analytic rep improve
This query improve the performance of the report in case there isn't any
archived analytic account.

Before this commit:
If there isn't any result in the `account_ids_to_not_display` list, the
query will be performed again.

After this commit:
The query is performed only once in each and every case.

closes odoo/enterprise#21744

X-original-commit: b12199178577cf88ef109933fb7d3ba274a88ee3
Signed-off-by: Florian Gilbert <FlorianGilbert@users.noreply.github.com>
Signed-off-by: Brice Bartoletti <Levizar@users.noreply.github.com>