[IMP] appointment{_crm}: link lead to appointment type
When an user creates a custom appointment type from the Meeting
tab in a lead, it should link the event that will be created
later to this particular lead. In this way, a CRM manager can
better view its meeting related to lead.

We add a new field to the appointment type model "opportunity_id"
that will be set based on the default_opportunity_id passed in the
context.

task-2665863

Part-of: odoo/enterprise#21799