[FIX] industry_fsm_sale: ensure product is a service one for timesheets
Before this commit, since the adding of new field in `product.template`
model called `detailed_type` to define the detailed type of a product.
We need to update a domainto select a product to timesheets on it when
the project is billable.

This commit updates the domain in `timesheet_product_id` field in
`project.sale.line.employee.map` model to be sure the product selected
is a service one and not a event or event booth product for instance.

task-2642872

Community PR: #79034

closes #21907

X-original-commit: 1f54dc9b37f624f369c8a8b7b1380d1cf9af4895
Part-of: odoo/enterprise#22056