[FIX] sale_planning: fix `is_service` not found in SOL model
Before this commit, we use the `is_service` field in
`sale_planning_admin_read_all_sol` ir.rule to have only SOL with product
type is equal to 'service'. However, this field is defined in
`sale_project` module. So, if the user has not this module installed this
ir.rule does not work.

This commit updates the `ir.rule` to replace the `('is_service', '=', True)`
into `('product_id.type', '=', 'service')`.

Steps to reproduce:
==================

1) Install sale_planning in empty db.
2) Log in as Mitchell Admin.
3) Go to the user form view of Marc Demo to change its access rights.
4) Give 'Administrator' access right in Planning app for this user and
remove the others access rights.
5) Log out and log in as 'Marc Demo'.
6) Go to planning app.
7) Select a shift and try to select a SOL for this shift.

Expected behavior:
=================

Some SOLs should be displayed to select one for this shift.

Actual behavior:
===============

Traceback because of the `is_service` field used in `ir.rule`.

task-2642872

Community PR: #79034

closes #21907

X-original-commit: d180c14af045cff20009ec6c62aaa12ceb1b3a2b
Part-of: odoo/enterprise#22056