[FIX] web_gantt: order by start date in gantt view
This PR purpose is to ensure that the records displayed in the gantt view
are chronogically displayed when being part of a same cell.

Let's say we have the following:

    - Task 1: sequence = 1, starts on 2021-02-04 and ends 2021-02-05
    - Task 2: sequence = 2, starts on 2021-02-03 and ends 2021-02-04
    - Task 3: sequence = 3, starts on 2021-02-01 and ends 2021-02-15

All tasks are assigned to you and the gantt view is in month mode over the period
feb 2021

Current behavior:

    - In the february cell of your user, you see, considering top to bottom order:

        - Task 1
        - Task 2
        - Task 3

Expected behavior:

    - In the february cell of your user, you should see, considering top to bottom order:

        - Task 3
        - Task 2
        - Task 1

Reason:
The search performed by the rpc returns records according to the order set on the model.
As such, they are treated in that order and this causeis the chronological order not to be respected
when pills are part of the same cell. At the time of this commit the order in task
is "priority desc, sequence, id desc" which is obviously not chronological.

task-2516896

closes odoo/enterprise#21691

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>