[FIX] hr_expense_extract: remove duplicate button in expenses settings
Before this commit the '-> View My Services' button appears two times in
the expenses settings view. The reason is the button is addded in the
`iap_buy_more_credits` widget when `hideService` options is false.

This commit hides the button in expense settings view to let the widget
added this button.

task-2665899

Community PR: #79034

closes #21907

closes odoo/enterprise#22056

X-original-commit: 9a6f2c7cf2fdac093a67ac87d62dcdff25b13b05
Related: odoo/odoo#79291
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>