[IMP] hr_appraisal: compute appraisal plan accurately
Before this commit, the date of the first appraisal (after recruitment)
is computed based on the create_date of the employee.

After this commit, if 'Employee Contracts' is installed, the date of the
first appraisal (after recruitment) is going to be based on the start of
the first contract (in case it's populated) for the employee.
Otherwise, create_date will be used.

The field next_appraisal_date is 'rebranded'.
Now it reflects the date at which an appraisal will be created by the appraisal plan.

task - 2667125

closes odoo/enterprise#21759

Signed-off-by: Kevin Baptiste <kba@odoo.com>