[IMP] mail_enterprise*: swiper component to perform actions
*: web_enterprise

This commit adds the ActionSwiper component. This component has
been made as a generic web_enterprise component for reusability.
Currently used in mail_enterprise, it allows mobile users to
swipe to right a message or conversation and mark it as read.
A swipe to the left unpins the conversation.
It can be added to other components if needed.

task-2391177
task-2361187

closes odoo/enterprise#21712

Related: odoo/odoo#78426
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>