[FIX] account_bank_statement_import_camt: no currency rate found
It seems banks don't have the same convention, some are putting journal's currency
in SrcCcy and others in TgtCcy, so we should inspect both indifferently.
In both cases though, the currency rate provided is to be used in the same way, we
only need to invert it in case the currency rate is provided in the counter value
element.

X-original-commit: 9ff5cca82bdf9c6c46d574fe04aea5b8a3c8fd62
Part-of: odoo/enterprise#22138