[FIX] account_bank_statement_import_camt: improve amount currency set
Before this commit, we were only looking in InstdAmt element to retrieve the
foreign currency and its related amount that are set on statement lines after.
We should also inspect TxAmt and Amt elements, as those can also be set with
a foreign currency. This will allow an easier reconciliation later as the
information about the foreign currency used is not lost and correctly imported.

closes odoo/enterprise#22138

X-original-commit: 1e85bf71d8f2e3c596b50e5d6a9ed7e452cee4e1
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>