[FIX] sale_subscription_dashboard: crash when rendering graph
Before this commit, there was a race condition in the dashboard
that could cause a crash. Sometimes, the call to the Chart.js lib
to render the chart could have been done when the dashboard wasn't
in the DOM yet. When this happenned, there was a crash in the lib
as it tries to compute positions by looking into the DOM.

This commit ensures that we wait to be inserted in the DOM before
rendering the chart.

This issue appears since [1] as with owl, the dashboard is inserted
in the DOM in the next animation frame (so a few ms after being
rendered).

[1] 29731b404fe624b733d784cd4321113d6b0d27b7

closes odoo/enterprise#21735

X-original-commit: c10261ab58dbc78c8b804bc42432d47324d77668
Related: odoo/odoo#78494
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>