[FIX] industry_fsm_report, quality_control_worksheet: batch demo data
To demonstrate the amazingness of worksheets the demo data of these
modules go and create a bunch of custom fields.

As a result these go through the entire flush/setup_models/init_models
once per field, which is pretty expensive.

Bulk-creating those custom fields should only require re-setting up
the models once for the entire batch skipping the vast majority of the
overhead. Worksheets create a custom model to which we have to attach
our custom fields, so CSVs are not an option.

Part-of: odoo/enterprise#21746