[FIX] pos_loyalty: allow apply multiple times fixed amount
When you are using a fixed amount reduction in the loyalty program, it
only applies one time the cost of point for this reward. To avoid
having to click multiple times on the available reward, we are spending
as many points as possible in the reward fixed amount and set it as
quantity.

This behavior consist of having a similar behavior as with the resale
reward type in v 13.0

Here is the example of the wanted behavior:
You want to give a reward of 10% based on what the customer purchased
previously. And you want to be able to apply it at any time, not just
waiting to have 5$ of reduction to give it to customer.

The configration made is this one.
On the program, you get 100 points/$
On the reward rule, it cost you 10 points to get a fixed amount of 0.01$
It means that based on what you get as points, you are giving 10%

Before this fix, if you sold for 1$ => get 100 points, you were abe to
give the reward as it cost 10 points. But you had to click 10 times, to
spende the 100 points available.

Now, it will apply directly 10 times the discount whent it is a fixed
amount.

closes odoo/enterprise#21794

X-original-commit: ce5b504125508fc08164ab48a10cde8320b9064e
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>