[IMP] web_studio: add performance test
Creating a model in studio on runbot with all modules installed enabling
all options takes more than 60 seconds and timeouts. Even with a few
options,the request last more than 10 seconds.

The main reason for this slowness is the number of call to
`setup_models`. This test shows the current number of call before
futures optimizations.

Part-of: odoo/enterprise#21746