[IMP] web_studio: batch models and fields creation
When creating a new model in studio, the initial logic
creates all models and fields individually.

The main problem is that each time a field or model is created, a call
to `setup_models` is made to adapt the registry to the model changes. On
a database witch a lot of modules installed, this can take around 3
seconds.

Before this changes, 51 calls to `setup_models` where triggered during a
call to `studio_model_create`, leading to a timeout. The estimate time to
finish the request would be around 2 minutes.

The new strategy is to accumulate all models and fields in order to
create everything at once, and only make one call to `setup_models`.

With this commit, `studio_model_create` only takes a few seconds.

closes odoo/enterprise#21746

Related: odoo/odoo#78514
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>