[IMP] hr_appraisal: add multi-edit to date_close
Purpose: Recent changes in Employee allow user to create a lot of
appraisals at once. But you can't edit anything on them on the same time.
It's not efficient if you can create them in a batch, but not edit some fields.

Thus, after this commit it is possible to multi-edit date_close.

task - 2669593

closes odoo/enterprise#21776

Signed-off-by: Kevin Baptiste <kba@odoo.com>