[MOV] documents_spreadsheet: move all the files into a reasonnable place
This is a preliminary step to be able to lazy load the entire o-spreadsheet feature in odoo.
As a lot of files have inter-dependencies, I am moving the files that will be lazy loaded to
they own module and reorganize the files left into a comprehensible structure.

The following commit will update all the references inside the moved files.

Part-of: odoo/enterprise#21804