[IMP] sale_subscription: Detect when an invoice is ready to be sent
In some flow like subscription, invoices are sent by mail automatically to the customer.
Sometimes, the invoice must be approved by the government before sending the mail to the customer like in the Mexican EDI.
This commit aims to add a custom method to detect when an invoice is ready to be sent to the customer.

PR (community): https://github.com/odoo/odoo/pull/78714
PR (enterprise): https://github.com/odoo/enterprise/pull/21809

closes odoo/enterprise#22171

X-original-commit: 9d31e2d3222dfd9d2ff9d1c948188601e67dfba4
Related: odoo/odoo#79504
Signed-off-by: William André (wan) <wan@odoo.com>