[FIX] web_grid: update grid_anchor in grid model when the range changes
Before this commit, when the user is in the timesheet grid view in
`week` range, he sees the current week. If he changes to the following
week and clicks on "add a line" button the default_date is 7 days after
the current one. This behavior is right but when we change the range to
`month` and the current day it is in the month displayed in the grid
view and the user clicks again on the `add a line` button. The
`default_date` is not the current date but 7 days after the current
date. The problem is we don't update the `grid_anchor` and `default_date`
in the context of the grid view when the range changed.

This commit updates the context after fetching the data for the new
range. That is, when the current range is `week` and the user clicks on
`month` button to have the grid view in `month` range. We check if the
current date is one of the new columns in the grid view. If it is the
case then the `grid_anchor` and `default_date` will be equal to the
current date.

I removed assert for th = 'Web Apocalyst' because no records for this
project are between 01/23/2017 and 01/29/2017. So normally, this group
should not be in the current range for this test.

Closes odoo/enterprise#21692
task-2652210

Related: odoo/odoo#80495
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Co-authored-by: ppr-odoo <ppr@odoo.com>