[IMP] stock_barcode: allow deleting of picking moves
Before this commit: opening a picking and scanning a product that isn't
part of the picking => no way to remove it from the picking. Workaround
was to set it to 0 quantity.

This commit allows users to remove this extra move line via the edit
(i.e. pencil button) view. Move lines that are part of the original
picking are still not deletable and inventory adjustment lines are
unaffected.

closes odoo/enterprise#21819

Task: 2658903
Signed-off-by: Steve Van Essche <svs@odoo.com>