[IMP] appointment: improve time zone picking on user interface
Purpose
=======
Improve the user experience with timezones when booking an appointment
from the website.
The timezone is set to the appointment timezone if the location is set
(since the user will probably have to go there in person), and to the
user timezone if location is not set.

Fix calendar days disappearing when selecting timezone.
Specifications
=============
Set timezone field to be always visible and editable (without having to activate
it in "Customize"). Change layout of selection field for time zone.

The default timezone is the one from set in the appointment if the appointment location is
set, otherwise it's the visitor's timezone.
Fix weekdays' names not being displayed when changing timezone.

Task-2672237

closes odoo/enterprise#21824

Related: odoo/upgrade#2954
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>