[FIX] project*: fix groups on computed field
*: project_account_asset,
   project_hr_payroll_account,
   project_sale_subscription

This commit fixes issue regarding fields in project form stat buttons.
Those fields are evaluated in any cases, and should rather be computed
only when the user has the right group.

Steps to reproduce :
1) Install project with one of this bridge
2) Connect with a user with only the group_project_admin right
3) Go to the form view
4) Issue:
An internal user with the group "Project Administrator" is unable to
access the form view of a project because of missing rights on the
X model.

Desired Behavior
================
The stat button should not be shown and the field not computed.

opw-2675514

closes odoo/enterprise#21889

X-original-commit: bc93fce6b022a0e5c31584f48a7de765cb769ada
Related: odoo/odoo#78981
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>