[IMP] account_sepa: smarter management of company name in SCT XML files
In SEPA Credit Transfer XML files, some elements are using the company name
(amongst other parameters that enforces unicity): MsgId, PmtInfId and
EndToEndId.

Before this commit, we were using the company name directly and crop it to
its last 15 characters. This was creating some confusion on banks' side as
some elements like EndToEndId are visible by end users.

After this commit, we now crop the company name to its first 15 characters
to make it easier to understand for the end user, as well as using the
creditor reference party name set on the company to allow some flexibility
in case some wants to use a shorter of different name in their files.

This commit also fixes a small bug when filling in the Nm element if the
initiating party name is not filled in by adding a fallback on the company
name instead.

closes odoo/enterprise#21969

X-original-commit: d911e980e918043f565288b56bf40bf2595455dc
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>