[IMP] project_enterprise: display allocated hours in popover
This commit adds the allocated hours information on the gantt popover.

The allocated hours are computed based on the user's working calendar
information.
If there are multiple assignees, the working hours is the sum of the
allocated hours for each user.

This commit takes advantage of the PR #21135 and copy the way the
allocated hours are computed in planning slots.

task-2664117

closes odoo/enterprise#21877

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>