[IMP] planning: avoid plan orders on multiple employees
Purpose
=======

Prior to this commit, when a sale order line was already planned during
a given period for a given employee, the next call to action_plan_orders
was still planning a new slot for another employee.
The goal of this task is to avoid planning same sale order line to
multiple employees within the same period.

Specification
=============

The employees assigned on sale line orders already planned during the
period are retrieved and then are the only employees assignable on those
sales order lines.

task-2654922

closes odoo/enterprise#21881

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>