[IMP] mail_enterprise,*: change model declaration syntax
 * = approvals, voip, website_helpdesk_livechat

This commit will change the declaration of the models with the aim of
getting closer to a plain JSON-object. It allows us to eliminate a wide
portion of boilerplate code, making the declaration shorter, more
declarative, and less redundant, but mainly, it prepares the ground for
further changes.

Task-2695223

closes odoo/enterprise#22042

Related: odoo/odoo#79259
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>