[FIX] account_accountant: reposition tooltip to avoid resize flicker
The tour bubble animation that makes it to bounce up and down can
cause issues when its position is at the edge of the bottom of the screen.
this would make the window constantly resize to show a scrollbar
and then resize to hide the scrollbar.

so this commit fixes the issue by changing the position of the tooltip
indicator from bottom to the top to avoid the screen glitch

TaskID-2667140

closes odoo/enterprise#21946

X-original-commit: eacebab7ab062d277220b51eeaad2daf6f090793
Related: odoo/odoo#79129
Signed-off-by: Laurent Smet <las@openerp.com>