[FIX] l10n_cl_edi: Take into account company id when matching on issuer VAT
The cronjob that fetches DTE mails runs under OdooBot, which has
superuser rights. This means that it ignores record rules and allowed
company ids from the current context. `_get_partner` was looking for the
contact of the issuer by searching purely on the VAT number. If there
was a contact for the same VAT defined on a different company, that
would otherwise be inaccessible to the company associated with the
invoice, it would get matched as a result of that query and running as
the superuser.

This had several consequences. First the creation of the invoice would
fail at the end of the `create()` function, throwing an exception:
https://github.com/odoo/odoo/blob/cd144298c82dacd2e1919b79dd1c2307b55da64b/odoo/models.py#L3909 .
Next that exception gets caught and a dummy record with no invoice lines
is created: https://github.com/odoo/enterprise/blob/db280068a16c6f29a25de47c40e39ce02e1ec749/l10n_cl_edi/models/fetchmail_server.py#L253 .
This leaves the user with two invoices. The first invoice will show
a read access error when opening it, because the user won't be able
to read the contact. The second invoice will be for the amount 0 and be
empty.

So there are two issues here: the query for the partner isn't specific
enough, and if an exception happens when creating the invoice, the
failed invoice persists.

This commit makes `_get_partner` take into account the company id so the
correct partner gets matched (or not if there is none). It will also
match on contacts that don't have a company set and are thus available
to all companies.

To prevent exceptions in the creation of the first or second invoice
from leaving behind bad records in combination with current exception
handling, both calls can be wrapped in savepoint contexts. Since this
isn't strictly necessary to fix this particular bug because the partner
matching is the root cause, this is left to be done in master instead.

opw-2633022

closes odoo/enterprise#22260

X-original-commit: 487ca0c9cb8ec70480d78661a9fd690fcc36b20f
Signed-off-by: Josse Colpaert <jco@openerp.com>