[FIX] sign: fixes signature wizard not opening in edit while signing mode
In edit while signing mode, it is possible to add signature sign items, move
them and finally sign them. Since the signature/initial sign item added while
signing can be moved/resized, the HTML structure is different from a normal
signature/initial sign item, therefore some edge cases need to be treated.

Currently, when you sign a added signature sign item, the first sign is done
when the user clicks in the div with class 'o_sign_item_display'. After
signing, the signature image takes the whole space and the move/resize handlers
are deleted. So if you want to click the sign item again to edit the signature,
the wizard doesn't open because we had registered the click event on the
'o_sign_item_display'. This PR changes the event listener to react to all
clicks on the sign item and then check if it is in editMode and if
'o_sign_item_display' is there. This fixes the wizard not opening for the
second time you click an added sign item trying to change the signature.

task-2679957

closes odoo/enterprise#22565

X-original-commit: 841ff81d9d2de2a906d97582f8e746d21cf35324
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>