[FIX] industry_fsm: get user_ids and not command list in default_get
Before this commit, we try to get the user_ids to check if only one user
is assigned to the task in result of the `default_get` method. The
problem is: `result.get('user_ids')` is a list of commands and not a
list of user ids.

This commit fixes this issue to get the user ids in the last
Command.SET containing in the commends list.

Steps to reproduce:
==================
1) Go to Field Service app,
2) Go to the gantt view,
3) Click on `Add` button.

Observed behavior:
=================
Traceback since the list is not iterable.

Expected behavior:
=================
Get the default planned dates for the task based on the user assigned.

closes odoo/enterprise#22093

X-original-commit: 9fb01d03b5ab27f448569650e477ffc341c59b15
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>