[REF] sign, documents_sign: refactor upload_template
move the logic `upload_template` to `create`
enable create_multi for template
(client still doesn't call `multi-create` to create templates in batch
for sake of stability)
use mimetypes.py to guess the mimetypes

The new relationship between attachment and sign.template while creating a sign.template
(A copy of the attachment is used to create the new sign.template if it has already linked to a non-'document.document' model)
* If a pdf is uploaded from Sign/or somewhere else (res_model == False).
The attachment will be directly used for the template. (unlink the template <=> unlink the attachment)
The attachment will be updated with (res_model == 'sign.template')
And we will not create a Document while creating the sign.template

* If a pdf is uploaded from Documents (res_model == 'document.document'), and you 'Sign' from the Documents.
The attachment will be directly used for the sign.template. (unlink the template <=> unlink the attachment => unlink the Document)
and the attachment will be updated with (res_model == 'sign.template')

* If a pdf is linked to an existing template or another model (res_model == 'sign.template'/'xxx').
A copied attachment will be used for the template.
(unlink the template <=> unlink the copied attachment, the original attachment and its linked record will not be affected)

Before this commit:
1.`upload_template` is the name for web client. The corresponding name for server is actually `create`
2. Both `upload_template` and `create` are used to create a new sign.template, which doesn't respect the Single responsibility principle.
3. mimetype is analyzed by js, which is not correct sometimes(pdf file without .pdf extension)
4. In sign the relationship between sign.template and its attachment is one 2 one, but Document doesn't respect this rule. As a result, sometimes deleting a sign.template/attachment will delete multiple related sign.templates

Taskid: 2659342
Part-of: odoo/enterprise#22813