[FIX] web: fix behavior of actionswiper
This commit fixes multiple issues with the ActionSwiper
component. Those changes improves its usage and reliability.
The first issue was due to the lack of rtl support when using
the component to perform actions because the actions were
not inverted as they should be. Now, instead of using props
directly, we use localizedProps that will change depending
of the language direction. A test has been added.

The second issue that has been fixed is about the swiper not
having the right height because of overflow issues. A div
now wraps the content to hide only the overflow-x property
without affecting the y axis.

Finally, the swiper now behaves correcly when scrollables areas
are under touch manipulation. Now, the element can be swiped
when all scrollable areas have reached their limit.
e.g. This bug was preventing interaction with emails formatted
with elements larger than the actual email. A test has been
written to assure that remain functionnal.

closes odoo/enterprise#22022

Related: odoo/odoo#79252
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>